[feat] : main.html 완성 및 자잘한 오류 수정

+++ src/ts/loginCheck.ts
@@ -11,8 +11,10 @@ function storageChk() {
 
 storageChk();
 // 로그아웃
-const logoutBtn = document.querySelector("#logoutBtn") as HTMLSpanElement;
-logoutBtn.onclick = (e: Event) => {
-    sessionStorage.clear();
-    location.reload();
+if (document.querySelector("#logoutBtn") !== null) {
+    const logoutBtn = document.querySelector("#logoutBtn") as HTMLSpanElement;
+    logoutBtn.onclick = (e: Event) => {
+        sessionStorage.clear();
+        location.reload();
+    }
 }
